+++ src/conventional/cum2x.py
@@ -58,7 +58,7 @@ def cum2x(x, y, maxlag=0, nsamp=0, overlap=0, flag='biased'):
     scale = make_arr((list(range(lx-maxlag, lx+1)), list(range(lx-1, lx-maxlag-1, -1))), axis=1).T
     scale = np.ones([2*maxlag+1, 1]) / scale
 
-  ind = np.arange(nsamp).T
+  ind = np.arange(nsamp, dtype=int).T
   for k in range(int(nrecs)):
     xs = x[ind].ravel(order='F')
     xs = xs - np.mean(xs)
